refactor!: Split visuals into their own crate
This significantly speeds up compile times when not using the visuals,
particularly notable on doctests.
Also good practice for the future
to make sure information needed for visuals is exposed from the main crate,
allowing potential users to make their own.
Plus it allows separate version releases for the two, which seems likely to happen.
Slight loss in dev ergonomics but definitely worth it.