refactor(vis): Move arrow cap to cover segment end; other minor adjustments
Decided to do this instead of implementing an offset to line segments
because that turned out to be really annoying to do
and my use cases don't really depend on the arrow point
being in an exact specific spot.

Also removed a case where circles were drawn twice
and moved instance geometry generating code
to a place where it's easier to find.