refactor(vis): Adapt to run on WebGL
Just needed to get rid of storage buffers and disable MSAA.
This changes no user-facing APIs and shouldn't break anything (besides MSAA),
although it would be nice to add the WebGPU platform back as an option
since anti-aliasing is a big deal for visual quality.

No guarantees about long-term support for WebGL,
but for now I can't see why we'd need to drop it.